pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lemmatizer"
version = "0.1.0"
description = "Arabic Lemmatization Pipeline using CAMeL Tools and LLM resolution"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "python-dotenv>=1.0.0",
    "pyarrow>=14.0.0",
    "camel-tools>=1.5.0",
    "requests>=2.31.0"
]

[tool.setuptools.packages.find]
where = ["src"]



